requirements.txt
RPi.GPIO==0.7.1
pandas==2.0.3
flask==2.2.2
requests==2.31.0
Adafruit-SSD1306==1.6.1
